test_Scripts/requirements_simple.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
chromadb==0.4.18
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-magic-bin==0.4.14
pydantic==2.5.0
python-dotenv==1.0.0
